requirements.txt
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
pyTelegramBotAPI==4.21.0
python-dotenv==1.0.1
python-telegram-bot==20.8
pytz==2024.1
requests==2.32.3
schedule==1.2.2
sniffio==1.3.0
telebot==0.0.5
urllib3==2.2.2
google-cloud-storage==2.10.0
google-cloud-scheduler==2.13.5
flask==2.3.2
